requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altgraph==0.17.4
attrs==24.2.0
discord==2.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.10
macholib==1.16.3
multidict==6.1.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
qasync==0.27.1
six==1.16.0
tzdata==2024.1
yarl==1.11.1
